making corrections
