+++ src/js/data.js
@@ -31,13 +31,12 @@ export const FILTER_DATA = [
 ];
 
 export const TASK_DATA = {
-  MIN: 1,
-  MAX: 7,
-  MIN_NUM_HESTAG: 0,
-  MAX_NUM_HESTAG: 3,
-  COLOR_CLASS: [`black`, `yellow`, `blue`, `green`, `pink`],
-  TYPE: [`repeat`, `deadline`, ``],
-
-  TITLE: [`Изучить теорию`, `Сделать домашку`, `Пройти интенсив на соточку`],
+  MIN_TASK_COUNT: 1,
+  MAX_TASK_COUNT: 7,
+  MIN_TAGS_COUNT: 0,
+  MAX_TAGS_COUNT: 3,
+  COLOR_CLASSES: [`black`, `yellow`, `blue`, `green`, `pink`],
+  TYPES: [`repeat`, `deadline`, ``],
+  TITLES: [`Изучить теорию`, `Сделать домашку`, `Пройти интенсив на соточку`],
   TAGS: [`homework`, `theory`, `practice`, `intensive`, `keks`],
 };
